Remove unnecessary ClipboardUtils class
